refactored packages for groupstage and knockoutstage
