+++ backend/src/main/java/org/fencing/demo/events/EventRepository.java
@@ -2,7 +2,7 @@ package org.fencing.demo.events;
 
 import java.util.List;
 
-import org.fencing.demo.stages.*;
+import org.fencing.demo.knockoutstage.KnockoutStage;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.query.Param;
