+++ backend/src/main/java/org/fencing/demo/events/EventServiceImpl.java
@@ -3,12 +3,12 @@ package org.fencing.demo.events;
 import java.util.List;
 
 import org.fencing.demo.exception.MatchesNotCompleteException;
+import org.fencing.demo.groupstage.GroupStage;
+import org.fencing.demo.knockoutstage.KnockoutStage;
 import org.fencing.demo.match.Match;
 import org.fencing.demo.player.Player;
 import org.fencing.demo.player.PlayerNotFoundException;
 import org.fencing.demo.player.PlayerRepository;
-import org.fencing.demo.stages.GroupStage;
-import org.fencing.demo.stages.KnockoutStage;
 import org.fencing.demo.tournament.TournamentNotFoundException;
 import org.fencing.demo.tournament.TournamentRepository;
 import org.springframework.stereotype.Service;
