+++ backend/src/main/java/org/fencing/demo/groupstage/GroupStage.java
@@ -1,4 +1,4 @@
-package org.fencing.demo.stages;
+package org.fencing.demo.groupstage;
 
 import java.util.List;
