+++ backend/src/main/java/org/fencing/demo/groupstage/GroupStageServiceImpl.java
@@ -1,4 +1,4 @@
-package org.fencing.demo.stages;
+package org.fencing.demo.groupstage;
 
 import java.util.ArrayList;
 import java.util.List;
