+++ backend/src/main/java/org/fencing/demo/knockoutstage/KnockoutStageServiceImpl.java
@@ -1,4 +1,4 @@
-package org.fencing.demo.stages;
+package org.fencing.demo.knockoutstage;
 
 import java.util.List;
