+++ backend/src/main/java/org/fencing/demo/match/MatchServiceImpl.java
@@ -9,14 +9,14 @@ import org.fencing.demo.events.Event;
 import org.fencing.demo.events.EventNotFoundException;
 import org.fencing.demo.events.EventRepository;
 import org.fencing.demo.events.PlayerRank;
+import org.fencing.demo.groupstage.GroupStageNotFoundException;
+import org.fencing.demo.groupstage.GroupStageRepository;
+import org.fencing.demo.knockoutstage.KnockoutStage;
+import org.fencing.demo.knockoutstage.KnockoutStageNotFoundException;
+import org.fencing.demo.knockoutstage.KnockoutStageRepository;
 import org.fencing.demo.player.Player;
 import org.fencing.demo.player.PlayerNotFoundException;
 import org.fencing.demo.player.PlayerRepository;
-import org.fencing.demo.stages.GroupStageNotFoundException;
-import org.fencing.demo.stages.GroupStageRepository;
-import org.fencing.demo.stages.KnockoutStage;
-import org.fencing.demo.stages.KnockoutStageNotFoundException;
-import org.fencing.demo.stages.KnockoutStageRepository;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
