+++ backend/src/main/java/org/fencing/demo/matchMaking/GroupMatchGenerator.java
@@ -7,8 +7,8 @@ import java.util.TreeMap;
 
 import org.fencing.demo.events.Event;
 import org.fencing.demo.events.PlayerRank;
+import org.fencing.demo.groupstage.GroupStage;
 import org.fencing.demo.match.Match;
-import org.fencing.demo.stages.GroupStage;
 import org.springframework.stereotype.Component;
 
 @Component
