+++ backend/src/test/java/org/fencing/demo/KnockoutStageIntegrationTest.java
@@ -13,11 +13,11 @@ import org.fencing.demo.events.Event;
 import org.fencing.demo.events.EventRepository;
 import org.fencing.demo.events.Gender;
 import org.fencing.demo.events.WeaponType;
+import org.fencing.demo.knockoutstage.KnockoutStage;
+import org.fencing.demo.knockoutstage.KnockoutStageRepository;
 import org.fencing.demo.match.Match;
 import org.fencing.demo.player.Player;
 import org.fencing.demo.player.PlayerRepository;
-import org.fencing.demo.stages.KnockoutStage;
-import org.fencing.demo.stages.KnockoutStageRepository;
 import org.fencing.demo.tournament.Tournament;
 import org.fencing.demo.tournament.TournamentRepository;
 import org.fencing.demo.user.Role;
