+++ backend/src/test/java/org/fencing/demo/KnockoutStageServiceTest.java
@@ -19,10 +19,10 @@ import org.fencing.demo.events.EventNotFoundException;
 import org.fencing.demo.events.EventRepository;
 import org.fencing.demo.events.Gender;
 import org.fencing.demo.events.WeaponType;
-import org.fencing.demo.stages.KnockoutStage;
-import org.fencing.demo.stages.KnockoutStageNotFoundException;
-import org.fencing.demo.stages.KnockoutStageRepository;
-import org.fencing.demo.stages.KnockoutStageServiceImpl;
+import org.fencing.demo.knockoutstage.KnockoutStage;
+import org.fencing.demo.knockoutstage.KnockoutStageNotFoundException;
+import org.fencing.demo.knockoutstage.KnockoutStageRepository;
+import org.fencing.demo.knockoutstage.KnockoutStageServiceImpl;
 import org.fencing.demo.tournament.Tournament;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
