+++ backend/src/test/java/org/fencing/demo/MatchServiceTest.java
@@ -25,16 +25,16 @@ import org.fencing.demo.events.Gender;
 import org.fencing.demo.events.PlayerRank;
 import org.fencing.demo.events.PlayerRankComparator;
 import org.fencing.demo.events.WeaponType;
+import org.fencing.demo.groupstage.GroupStageRepository;
+import org.fencing.demo.knockoutstage.KnockoutStage;
+import org.fencing.demo.knockoutstage.KnockoutStageNotFoundException;
+import org.fencing.demo.knockoutstage.KnockoutStageRepository;
 import org.fencing.demo.match.Match;
 import org.fencing.demo.match.MatchNotFoundException;
 import org.fencing.demo.match.MatchRepository;
 import org.fencing.demo.match.MatchServiceImpl;
 import org.fencing.demo.player.Player;
 import org.fencing.demo.player.PlayerRepository;
-import org.fencing.demo.stages.GroupStageRepository;
-import org.fencing.demo.stages.KnockoutStage;
-import org.fencing.demo.stages.KnockoutStageNotFoundException;
-import org.fencing.demo.stages.KnockoutStageRepository;
 import org.fencing.demo.tournament.Tournament;
 import org.fencing.demo.tournament.TournamentRepository;
 import org.junit.jupiter.api.Test;
